add tooltip to some more component types
