+++ src/admin/ui/shared/LongTextField.tsx
@@ -1,25 +1,7 @@
 import React from 'react';
 import TextField from '@mui/material/TextField';
-import Tooltip from '@mui/material/Tooltip';
-import Icon from '@mui/material/Icon';
-import Box from '@mui/material/Box';
 import { Nullable } from 'game/store/types';
-import Help from '@mui/icons-material/Help';
-import { makeStyles } from '@mui/styles';
-
-const useStyles = makeStyles({
-  tooltip: {
-    position: 'absolute',
-    top: '2px',
-    right: '10px',
-    cursor: 'pointer',
-    padding: '2px',
-    backgroundColor: 'white',
-  },
-  icon: {
-    height: '20px',
-  },
-});
+import WithTooltip from './WithTooltip';
 
 type Props = {
   label: string;
@@ -31,9 +13,8 @@ type Props = {
 const LongTextField = ({
   label, value, onChange, fullWidth = true, tooltip,
 }: Props) => {
-  const classes = useStyles();
   return (
-    <Box style={{ position: 'relative' }}>
+    <WithTooltip text={tooltip}>
       <TextField
         label={label}
         margin="normal"
@@ -50,16 +31,7 @@ const LongTextField = ({
         }}
         variant="outlined"
       />
-      {tooltip && (
-        <Tooltip title={tooltip}>
-          <Box className={classes.tooltip}>
-            <Icon>
-              <Help fontSize="small" />
-            </Icon>
-          </Box>
-        </Tooltip>
-      )}
-    </Box>
+    </WithTooltip>
   );
 };
