+++ src/admin/ui/shared/assets/VideoSelector.tsx
@@ -6,6 +6,7 @@ import getFilenames from 'shared/util/getFilenames';
 import { Stack } from '@mui/material';
 import Selector, { makeOptions } from '../Selector';
 import VideoUploader from './VideoUploader';
+import WithTooltip from '../WithTooltip';
 // import VideoUploader from './VideoUploader';
 
 const loadVideos = async (gameName: string) => {
@@ -20,8 +21,11 @@ type Props = {
   label: string;
   value: Nullable<string>;
   onChange: (value: string) => void;
+  tooltip?: string;
 };
-const VideoSelector = ({ label, value, onChange }: Props) => {
+const VideoSelector = ({
+  label, value, onChange, tooltip,
+}: Props) => {
   const { gameName } = useParams<{ gameName: string }>();
   const [options, setOptions] = useState<Nullable<string[]>>(null);
 
@@ -40,13 +44,14 @@ const VideoSelector = ({ label, value, onChange }: Props) => {
 
   return (
     <Stack direction="row" spacing={2}>
-      <Selector
-        label={label}
-        value={options.length ? (value || '') : ''}
-        onChange={onChange}
-        options={makeOptions(options)}
-      />
-
+      <WithTooltip text={tooltip}>
+        <Selector
+          label={label}
+          value={options.length ? (value || '') : ''}
+          onChange={onChange}
+          options={makeOptions(options)}
+        />
+      </WithTooltip>
       <VideoUploader
         onSuccess={handleUploadSuccess}
       />
